improve layout; quizes are innerHTML
we can also have blank lines in quizes files and can comment with //

also added an abort controller. if fetching was not done and you would
switch to a different quiz, the process related to the previous quiz
might effect the state.

@va9iff will be always visible if answers are pulled down. max height of
both asking (the question) and answers are aligned to make a thin line
which will show the @va9iff even if they're overflowing

fix quizes list are scrollable on full screen too (might be previous
commit)

automatically pull down on mobile